chore: 공용 컴포넌트의 폴더 및 파일 구조 변경
