+++ src/common/components/BaseButton.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { IBaseButtonStyleProps } from '@/common/types/children.interface';
+
+export default function BaseButton({ className, children }: IBaseButtonStyleProps) {
+	return <button className={className}>{children}</button>;
+}
