+++ src/common/components/CloseButton.tsx
@@ -1,0 +1,5 @@
+export default function CloseButton() {
+	return (
+		<div className="bg-lightGray flex size-8 items-center justify-center rounded-full">Ⅹ</div>
+	);
+}
