--- src/common/components/CloseButton/index.tsx
@@ -1,5 +1,0 @@
-export default function CloseButton() {
-	return (
-		<div className="bg-lightGray flex h-8 w-8 items-center justify-center rounded-full">Ⅹ</div>
-	);
-}
