+++ src/common/components/ProfileEditButton.tsx
@@ -1,0 +1,5 @@
+import { IBaseButtonStyleProps } from '@/common/types/children.interface';
+
+export default function ProfileEditButton({ className, children }: IBaseButtonStyleProps) {
+	return <button className={className}>{children}</button>;
+}
